fix criteria input methods; refactor pairwise class; db add column; add profile abbr

+++ backend/app/db/models/criterion.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, Numeric
 from sqlalchemy.orm import Relationship
 from sqlalchemy.dialects import postgresql as pg
 
@@ -11,7 +11,7 @@ class Criterion(TimeStampedModel):
     criterion_id = Column(Integer, primary_key=True, autoincrement=True)
     criterion_name = Column(String(80), nullable=False)
     min_max = Column(String(3), nullable=False)
-    alternatives_values = Column(pg.ARRAY(Integer))
+    alternatives_values = Column(pg.ARRAY(Numeric))
     description = Column(String(500))
     task_id = Column(Integer, ForeignKey("tasks.task_id", ondelete="Cascade"), nullable = False, index = True)
